// angular-src/src/app/models/partnerlandingpage.ts
export class PartnerLandingPage {
    over30Count: number;
    over30Value: number;
    over60Count: number;
    over60Value: number;
    over90Count: number;
    over90Value: number;
    over120Count: number;
    over120Value: number;
    invoicesInProcess: number;
    invoicesInProcessValue: number;
    invoicesOnHold: number;
    invoicesOnHoldValue: number;
    invoicesApproved30days: number;
    invoicesApproved30daysValue: number;
    invoicesPaid30days: number;
    invoicesPaid30daysValue: number;
    pendingProspects: number;
    oldestProspect: Date;
    pendingCancels: number;
    rmrCancels: number;
    pendingCancels7days: number;
    highRMRCancelPerson: string;
    highRMRCancelValue: number;
    serviceTicketsPending: number;
    oldestInProgress: Date;
    attCount: number;
    verizonCount: number;
    sprintCount: number; //not used currently
    tMobileCount: number; //not used currently
    unknownCount: number;
    upgradesLastMonth: number;
    lifeTimeUpgrades: number;
    progressPercent: number;
    attritionLastMonth: number;
    attritionLast6Months: number;
    attritionLast12Months: number
}